Don't set branch in recipes if it is the default branch
For `insfactor' "release" isn't actually the default branch but it
points at the same commit as "master", which is the default branch,
and it is unlikely that they are going to diverge in the future as
the newest commit is eight years old.

Same for `neato-graph-bar'; though here it is only four years.